chore: remove IdentifierForm from Storybook, keep the component
IdentifierForm is not used by any route — the live classic identifier
screen renders IdentifierPage — but it remains a published public API
export, so the component itself stays. This only clears it out of
Storybook, where its presence alongside IdentifierPage was redundant.

- Delete the dedicated IdentifierForm.stories.tsx.
- AuthLayout.stories.tsx: swap the fixture to IdentifierPage across all
  layout stories. They demonstrate the layout, not the form, so they now
  render the live component. IdentifierPage requires non-optional
  theme/branding (IdentifierForm had them optional), hence the `?? null`
  on the loosely-typed story args — behaviourally identical.
- Port the two GitHub/Microsoft social stories to IdentifierPage.stories
  so the preview docs still have a target.
- Repoint packages/authhero/docs/github-microsoft-setup.md accordingly.

Stories are typechecked, not covered by Vitest (Storybook has its own
runner), so validated via tsc.
